+++ AtchI/HealthKitTemp/Model/HealthKitQuentitymodel.swift
@@ -1,0 +1,13 @@
+//
+//  HealthKitQuentitymodel.swift
+//  AtchI
+//
+//  Created by DOYEON LEE on 2023/04/17.
+//
+
+import Foundation
+
+struct HealthKitQuentityModel: HealthKitModel {
+    let identifier: HealthKitIdentifier
+    let quentity: Double?
+}
