+++ AtchI/HealthKitTemp/Model/Protocol/HealthkitDateModel.swift
@@ -1,0 +1,12 @@
+//
+//  HealthkitDateModel.swift
+//  AtchI
+//
+//  Created by DOYEON LEE on 2023/04/17.
+//
+
+import Foundation
+
+protocol HealthKitModel {
+    var identifier: HealthKitIdentifier { get }
+}
